+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+### Unreleased
+
+### Python Versions
+
+* Removed Python 3.7 support ([#376])
+* Added Python 3.12 support ([#376])
+
+### Maintenance
+
+* Migrate to a hatch-based build ([#374])
+
+### Fixed
+
+* Stop using Python's undocumented parse_template re function ([#378])
+* Resolved newline and encoding issues for running tests on Windows ([#379])
+
+
 ## [v1.8.1]
 
 **Release date: 2023-05-29**
@@ -1602,3 +1619,7 @@ information about changes, except for
 [#360]: https://github.com/delph-in/pydelphin/issues/360
 [#364]: https://github.com/delph-in/pydelphin/issues/364
 [#367]: https://github.com/delph-in/pydelphin/issues/367
+[#374]: https://github.com/delph-in/pydelphin/issues/374
+[#376]: https://github.com/delph-in/pydelphin/issues/376
+[#378]: https://github.com/delph-in/pydelphin/issues/378
+[#379]: https://github.com/delph-in/pydelphin/issues/379